Rename pkg/metrics/policyfilter to pkg/metrics/policyfiltermetrics
The previous package name included a typo ("policfilter") and the package was
imported as "pfmetrics" everywhere. This was confusing and inconsistent with
all other metrics packages. This patch renames it to "policyfiltermetrics" and
uses this name consistently.

Technically it is a breaking change, but only developer facing.
